First step of creating universal accessors for all objects that end up in the view
