+++ app/views/Search/object.html
@@ -9,7 +9,7 @@
 <div class="row">
     <div class="g-12of12 grid_12">
         <div class="media record">
-            <img src="${fullDoc.getThumbnail(100)}" class="img" width="350" alt="${fullDoc.getTitle()}"/>
+            <img src="${fullDoc.getThumbnailUri(350)}" class="img" width="350" alt="${fullDoc.getTitle()}"/>
             <div class="extra">
                 <a href="#" class="overlay-trigger" rel="#overlay" title="&{'ui.label.enlargeImage'}">&{'ui.label.enlargeImage'}</a>
             </div>
@@ -56,7 +56,7 @@
 <!-- overlayed element -->
 <div id="overlay" class="result-overlay" style="padding: 20px;width:auto">
 <div class="contentWrap">
-    <img src="${fullDoc.getThumbnail(100)}" alt="${fullDoc.getTitle()}"/>
+    <img src="${fullDoc.getThumbnailUri(500)}" alt="${fullDoc.getTitle()}"/>
     <div class="title"> ${fullDoc.getTitle()}</div>
 </div>
 </div>
@@ -66,24 +66,3 @@
         $("a.overlay-trigger").overlay();
 });
 </script>
-*{
-Possible methods
-def getAsArray(key: String) : Array[String] = assign(key)
-def getAsString(key: String) : String = assignFirst(key)
-def getFieldValue(key : String) : FieldValue = FieldValue(key, solrDocument)
-def getFieldValueList: JList[FieldValue] = solrDocument.getFieldValueList
-def getFieldValuesFiltered(include: Boolean, fields: Array[String]) : JList[FieldValue] = solrDocument.getFieldValuesFiltered(include, fields.toList)
-def getConcatenatedArray(key: String, fields: Array[String]) : FieldFormatted = solrDocument.getConcatenatedArray(key, fields.toList)
-
-def getId : String = assign("delving_hubId")
-def getDelvingId : String = assign("delving_pmhId")
-def getTitle : String = assign("title")
-def getDescription: String = assign("dc_description")
-def getThumbnail : String = assign("europeana_object")
-def getCreator : String = assign("dc_creator")
-def getYear : String = assign("europeana_year")
-def getProvider : String = assign("europeana_provider")
-def getDataProvider : String = assign("europeana_dataProvider")
-def getLanguage : String = assign("europeana_language")
-
-}*
