+++ app/views/Search/overlay.html
@@ -8,7 +8,7 @@
 <div class="row">
     <div class="g-12of12">
         <div class="media record">
-            <img src="${fullDoc.getThumbnail()}" class="img" alt="${fullDoc.getTitle()}"/>
+            <img src="${fullDoc.getThumbnailUri()}" class="img" alt="${fullDoc.getTitle()}"/>
 
             <div class="bd">
                 <div class="metadata">
