+++ app/views/tags/resultList.html
@@ -4,7 +4,7 @@
 <div class="row">
     <div class="media thing">
         <!-- TODO: get url to item page -->
-        <a href="#" class="clip"><img src="${doc.getThumbnail()}" class="img"/></a>
+        <a href="#" class="clip"><img src="${doc.getThumbnailUri()}" class="img"/></a>
         <div class="bd">
             <div class="mod">
                 <div class="inner">
